requirements.txt
git+https://gitlab.cern.ch/digitalmemory/bagit-create@develop
Django==5.0.6
djangorestframework==3.14.0
pymarc==4.0.0
redis==3.5.3
django-cors-headers==4.0.0
mozilla-django-oidc==3.0.0
psycopg2==2.9.6
sentry-sdk==1.3.1
oais-utils==0.0.4
django-celery-beat==2.6.0
amclient==1.2.0
celery==5.2.7
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
django-guardian==2.4.0
django-opensearch-dsl==0.2.0
importlib_metadata==4.11.3
fts3==3.12.0.post1

